pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clear>=2.0.0",
    "edge-tts>=7.2.3",
    "google-generativeai>=0.8.5",
    "gtts>=2.5.4",
    "nest-asyncio>=1.6.0",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pyaudio>=0.2.14",
    "pygame>=2.6.1",
    "speechrecognition>=3.14.4",
]
